refactor: merge query with projection
