+++ crates/contact/src/lib.rs
@@ -9,11 +9,9 @@ pub use types::*;
 cfg_if::cfg_if! {
     if #[cfg(feature = "full")] {
         mod command;
-        mod projection;
         mod query;
 
         pub use command::*;
-        pub use projection::*;
         pub use query::*;
     }
 }
