+++ crates/contact/src/query/mod.rs
@@ -1,0 +1,8 @@
+mod list;
+mod stat;
+
+pub use list::*;
+pub use stat::*;
+
+#[derive(Clone)]
+pub struct Query(pub sqlx::SqlitePool);
