+++ crates/contact/src/query/stat.rs
@@ -4,6 +4,35 @@ use imkitchen_db::table::ContactStat;
 use imkitchen_shared::Event;
 use sea_query::{Expr, ExprTrait, OnConflict, Query, SqliteQueryBuilder};
 use sea_query_sqlx::SqlxBinder;
+use sqlx::prelude::FromRow;
+
+#[derive(Default, Debug, FromRow)]
+pub struct Stat {
+    pub total: u32,
+    pub unread: u32,
+    // pub today: u32,
+    pub avg_response_time: u32,
+    // pub avg_response_time_last_week: u8,
+}
+
+impl super::Query {
+    pub async fn find_stat(&self, day: u64) -> anyhow::Result<Option<Stat>> {
+        let statement = sea_query::Query::select()
+            .columns([
+                ContactStat::Total,
+                ContactStat::Unread,
+                ContactStat::AvgResponseTime,
+            ])
+            .from(ContactStat::Table)
+            .and_where(Expr::col(ContactStat::Day).eq(day))
+            .to_owned();
+
+        let (sql, values) = statement.build_sqlx(SqliteQueryBuilder);
+        Ok(sqlx::query_as_with::<_, Stat, _>(&sql, values)
+            .fetch_optional(&self.0)
+            .await?)
+    }
+}
 
 pub fn subscribe_stat<E: Executor + Clone>() -> SubscribeBuilder<E> {
     evento::subscribe("contact-stat")
