+++ crates/db/src/m0_12/mod.rs
@@ -1,5 +1,5 @@
-mod mealplan_shopping_list;
 mod mealplan_slot;
+mod shopping_list;
 
 use sqlx_migrator::vec_box;
 
@@ -10,8 +10,5 @@ sqlx_migrator::sqlite_migration!(
     "imkitchen",
     "m0_12",
     vec_box![crate::m0_11::Migration],
-    vec_box![
-        mealplan_slot::CreateTable,
-        mealplan_shopping_list::CreateTable
-    ]
+    vec_box![mealplan_slot::CreateTable, shopping_list::CreateTable]
 );
