+++ crates/mealplan/src/lib.rs
@@ -1,6 +1,5 @@
 mod aggregator;
 mod event;
-pub mod shopping_list;
 mod types;
 
 pub use aggregator::*;
@@ -10,13 +9,11 @@ pub use types::*;
 cfg_if::cfg_if! {
     if #[cfg(feature = "full")] {
         mod command;
-        mod projection;
         mod query;
         mod service;
         // mod scheduler;
 
         pub use command::*;
-        pub use projection::*;
         pub use query::*;
         pub use service::*;
         // pub use scheduler::*;
