--- crates/mealplan/src/projection/mod.rs
@@ -1,7 +1,0 @@
-mod shopping_list;
-mod slot;
-mod week;
-
-pub use shopping_list::*;
-pub use slot::*;
-pub use week::*;
