+++ crates/mealplan/src/query/mod.rs
@@ -1,0 +1,8 @@
+mod slot;
+mod week;
+
+pub use slot::*;
+pub use week::*;
+
+#[derive(Clone)]
+pub struct Query(pub sqlx::SqlitePool);
