+++ crates/recipe/src/query/mod.rs
@@ -3,3 +3,6 @@ mod user_stat;
 
 pub use list::*;
 pub use user_stat::*;
+
+#[derive(Clone)]
+pub struct Query(pub sqlx::SqlitePool);
