+++ crates/shopping/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "imkitchen-shopping"
+version = "0.12.1"
+edition.workspace = true
+authors.workspace = true
+
+[dependencies]
+evento.workspace = true
+bincode.workspace = true
+anyhow.workspace = true
+sqlx = { workspace = true, optional = true }
+time = { workspace = true, optional = true }
+cfg-if.workspace = true
+sea-query.workspace = true
+sea-query-sqlx.workspace = true
+imkitchen-shared = { path = "../shared", version = "0.12.1" }
+imkitchen-db = { path = "../db", version = "0.12.1", optional = true }
+imkitchen-recipe = { path = "../recipe", version = "0.12.1", default-features = false, optional = true }
+imkitchen-mealplan = { path = "../mealplan", version = "0.12.1", default-features = false, features = ["imkitchen-recipe"], optional = true }
+
+[dev-dependencies]
+tokio.workspace = true
+temp-dir.workspace = true
+futures.workspace = true
+time = { workspace = true, features = ["formatting", "macros"] }
+
+[features]
+default = ["full", "imkitchen-recipe", "imkitchen-mealplan"]
+full = ["sqlx", "imkitchen-db", "time"]
