+++ crates/shopping/src/event.rs
