+++ crates/shopping/src/lib.rs
@@ -7,7 +7,9 @@ pub use event::*;
 cfg_if::cfg_if! {
     if #[cfg(feature = "full")] {
         mod command;
+        mod query;
 
         pub use command::*;
+        pub use query::*;
     }
 }
