+++ crates/shopping/src/query/mod.rs
@@ -1,0 +1,6 @@
+mod list;
+
+pub use list::*;
+
+#[derive(Clone)]
+pub struct Query(pub sqlx::SqlitePool);
