+++ crates/user/src/lib.rs
@@ -12,11 +12,9 @@ pub use types::*;
 cfg_if::cfg_if! {
     if #[cfg(feature = "full")] {
         mod command;
-        mod projection;
         mod query;
 
         pub use command::*;
-        pub use projection::*;
         pub use query::*;
     }
 }
