--- crates/user/src/projection/mod.rs
@@ -1,5 +1,0 @@
-mod list;
-mod stat;
-
-pub use list::*;
-pub use stat::*;
