+++ src/cli/server.rs
@@ -36,8 +36,8 @@ pub async fn serve(
     let recipe_query = imkitchen_recipe::Query(read_pool.clone());
     let mealplan_command = imkitchen_mealplan::Command(evento_executor.clone(), read_pool.clone());
     let mealplan_query = imkitchen_mealplan::Query(read_pool.clone());
-    let shopping_list_command =
-        imkitchen_mealplan::shopping_list::Command(evento_executor.clone(), read_pool.clone());
+    let shopping_command = imkitchen_shopping::Command(evento_executor.clone(), read_pool.clone());
+    let shopping_query = imkitchen_shopping::Query(read_pool.clone());
 
     // Start background notification worker
     tracing::info!("Starting evento subscriptions...");
@@ -93,12 +93,12 @@ pub async fn serve(
         .run(&evento_executor)
         .await?;
 
-    let sub_mealplan_shopping_list = imkitchen_mealplan::subscribe_shopping_list()
+    let sub_shopping_list = imkitchen_shopping::subscribe_list()
         .data(write_pool.clone())
         .run(&evento_executor)
         .await?;
 
-    let sub_shopping_list_command = imkitchen_mealplan::shopping_list::subscribe_command()
+    let sub_shopping_command = imkitchen_shopping::subscribe_command()
         .data(write_pool.clone())
         .run(&evento_executor)
         .await?;
@@ -118,7 +118,8 @@ pub async fn serve(
         recipe_query,
         mealplan_command,
         mealplan_query,
-        shopping_list_command,
+        shopping_command,
+        shopping_query,
         pool: read_pool.clone(),
     };
 
@@ -192,8 +193,8 @@ pub async fn serve(
         sub_mealplan_command.shutdown_and_wait(),
         sub_mealplan_week.shutdown_and_wait(),
         sub_mealplan_slot.shutdown_and_wait(),
-        sub_mealplan_shopping_list.shutdown_and_wait(),
-        sub_shopping_list_command.shutdown_and_wait(),
+        sub_shopping_list.shutdown_and_wait(),
+        sub_shopping_command.shutdown_and_wait(),
     ])
     .await;
