+++ templates/calendar.html
@@ -16,7 +16,7 @@
             <!-- Mobile: Stack vertically -->
             <div class="flex flex-col gap-3 md:hidden">
                 <div class="flex items-center justify-between gap-2">
-                    <a href="/calendar/week-{{ index }}/shopping-list"
+                    <a href="/calendar/week-{{ index }}/shopping"
                       class="px-3 py-2 bg-gray-200 text-gray-700 rounded-lg hover:bg-gray-300 text-sm">
                         🛒
                     </a>
@@ -43,7 +43,7 @@
 
             <!-- Desktop: Original layout -->
             <div class="hidden md:flex items-center justify-between">
-                <a href="/calendar/week-{{ index }}/shopping-list"
+                <a href="/calendar/week-{{ index }}/shopping"
                   class="px-4 py-2 bg-gray-200 text-gray-700 rounded-lg hover:bg-gray-300">
                     🛒 {{ "Shopping List"|t }}
                 </a>
