+++ templates/dashboard.html
@@ -124,7 +124,7 @@
                         <!-- <a href="/kitchen" class="block w-full px-4 py-3 bg-green-600 text-white font-semibold rounded-lg hover:bg-green-700 transition text-center"> -->
                         <!--     🍛 My kitchen -->
                         <!-- </a> -->
-                        <a href="/calendar/week-1/shopping-list" class="block w-full px-4 py-3 bg-blue-600 text-white font-semibold rounded-lg hover:bg-blue-700 transition text-center">
+                        <a href="/calendar/week-1/shopping" class="block w-full px-4 py-3 bg-blue-600 text-white font-semibold rounded-lg hover:bg-blue-700 transition text-center">
                             🛒 {{ "Shopping List"|t }}
                         </a>
                         <a href="/recipes/import" class="block w-full px-4 py-3 bg-gray-200 text-gray-700 font-semibold rounded-lg hover:bg-gray-300 transition text-center">
